Organize files in a more structured manner and fixes
