+++ src/app/auth/checkin/checkin.component.html
