+++ src/app/auth/checkin/checkin.component.ts
