+++ src/app/models/user.model.ts
@@ -1,0 +1,14 @@
+export interface IUser {
+    id: number;
+    fullName:string;
+    idNumber: number;
+    lastSeenOn: string;
+    password: string;
+    phoneNumber: string;
+    pic: string;
+    _id: string;
+    roles: [];
+    email: string;
+    accessToken: string;
+    address: {};
+}
