+++ src/app/modules/land-registrar/actions/land-owner-reg.component.ts
