+++ src/app/navigation/navbar.component.html
@@ -159,7 +159,7 @@
         <div class="kt-header__topbar-wrapper" data-toggle="dropdown" data-offset="10px,0px">
           <div class="kt-header__topbar-user">
             <span class="kt-hidden kt-header__topbar-welcome kt-hidden-mobile">Hi,</span>
-            <span class="kt-hidden kt-header__topbar-username kt-hidden-mobile">{{auth.currentUser.firstName}}</span>
+            <span class="kt-hidden kt-header__topbar-username kt-hidden-mobile">{{auth.currentUser.fullName}}</span>
             <img alt="Pic" class="kt-radius-100" src="./assets/media/users/300_25.jpg" />
   
             <!--use below badge element instead the user avatar to display username's first letter(remove kt-hidden class to display it) -->
@@ -177,7 +177,7 @@
               <span class="kt-badge kt-badge--username kt-badge--unified-success kt-badge--lg kt-badge--rounded kt-badge--bold kt-hidden">S</span>
             </div>
             <div class="kt-user-card__name">
-              {{auth.currentUser.firstName}} {{auth.currentUser.lastName}}
+              {{auth.currentUser.fullName}}
             </div>
             <div class="kt-user-card__badge">
               <span class="btn btn-label-primary btn-sm btn-bold btn-font-md">23 messages</span>
