--- src/app/user/user.model.ts
@@ -1,6 +1,0 @@
-export interface IUser {
-    id: number;
-    firstName:string;
-    lastName:string;
-    userName: string;
-}
